refactor: 요금을 분리하여 요금 계산 로직 이동

+++ src/main/java/wooteco/subway/domain/path/AgeDiscountStrategy.java
@@ -27,7 +27,7 @@ public enum AgeDiscountStrategy {
                 .orElseThrow(() -> new DataNotExistException("존재하지 않는 나이입니다."));
     }
 
-    public int discount(double fare) {
+    public int discount(int fare) {
         return (int) discountStrategy.applyAsDouble(fare);
     }
 }
